requirements.txt
streamlit
perplexityai
google-cloud-vision